refactor(cli): extract shared logDeliveryTime helper and add debug logging
Duplicate logDeliveryTime/logMessageDeliveryTime functions were deduplicated
into utils/relay.ts. The catch block now logs a debug message instead of
silently swallowing errors. Changeset updated to mention both warp send and
send message commands.
