+++ .changeset/nervous-mirrors-promise.md
@@ -2,4 +2,4 @@
 "@hyperlane-xyz/cli": minor
 ---
 
-Added transfer time for warp send transfers
+Delivery time logging was added to `hyperlane warp send` and `hyperlane send message`, displaying elapsed seconds between dispatch and delivery for EVM chains.
