+++ typescript/cli/src/send/message.ts
@@ -1,13 +1,8 @@
 import { stringify as yamlStringify } from 'yaml';
 
-import { Mailbox__factory } from '@hyperlane-xyz/core';
 import { GasAction } from '@hyperlane-xyz/provider-sdk';
 import { HyperlaneRelayer } from '@hyperlane-xyz/relayer';
-import {
-  type ChainName,
-  HyperlaneCore,
-  type MultiProvider,
-} from '@hyperlane-xyz/sdk';
+import { type ChainName, HyperlaneCore } from '@hyperlane-xyz/sdk';
 import { addressToBytes32, isEVMLike, timeout } from '@hyperlane-xyz/utils';
 
 import { EXPLORER_URL } from '../consts.js';
@@ -23,7 +18,7 @@ import {
   runSingleChainSelectionStep,
 } from '../utils/chains.js';
 import { indentYamlOrJson } from '../utils/files.js';
-import { stubMerkleTreeConfig } from '../utils/relay.js';
+import { logDeliveryTime, stubMerkleTreeConfig } from '../utils/relay.js';
 
 export async function sendTestMessage({
   context,
@@ -164,7 +159,7 @@ async function executeDelivery({
       log('Attempting self-relay of message');
       await relayer.relayMessage(dispatchTx);
       logGreen('Message was self-relayed!');
-      await logMessageDeliveryTime(
+      await logDeliveryTime(
         origin,
         destination,
         dispatchTx.blockNumber,
@@ -181,7 +176,7 @@ async function executeDelivery({
       // Max wait 10 minutes
       await core.waitForMessageProcessed(dispatchTx, 10000, 60);
       logGreen('Message was delivered!');
-      await logMessageDeliveryTime(
+      await logDeliveryTime(
         origin,
         destination,
         dispatchTx.blockNumber,
@@ -195,42 +190,5 @@ async function executeDelivery({
       `Encountered error sending message from ${origin} to ${destination}`,
     );
     throw e;
-  }
-}
-
-async function logMessageDeliveryTime(
-  origin: ChainName,
-  destination: ChainName,
-  dispatchBlockNumber: number,
-  messageId: string,
-  chainAddresses: Record<string, Record<string, string>>,
-  multiProvider: MultiProvider,
-): Promise<void> {
-  if (
-    !isEVMLike(multiProvider.getProtocol(origin)) ||
-    !isEVMLike(multiProvider.getProtocol(destination))
-  ) {
-    return;
-  }
-  try {
-    const mailboxAddress = chainAddresses[destination]?.mailbox;
-    if (!mailboxAddress) return;
-
-    const mailbox = Mailbox__factory.connect(
-      mailboxAddress,
-      multiProvider.getProvider(destination),
-    );
-    const processedBlockNum = await mailbox.processedAt(messageId);
-    const [dispatchBlock, processedBlock] = await Promise.all([
-      multiProvider.getProvider(origin).getBlock(dispatchBlockNumber),
-      multiProvider.getProvider(destination).getBlock(processedBlockNum),
-    ]);
-    if (dispatchBlock && processedBlock) {
-      logGreen(
-        `Delivery time: ${processedBlock.timestamp - dispatchBlock.timestamp}s`,
-      );
-    }
-  } catch {
-    // Non-fatal: delivery time is informational only
   }
 }
